requirements/req.txt
 flask
 mysql-connector-python